requirements.txt
certifi
charset-normalizer
future
idna
numpy
opencv-python
requests
termcolor
tk
urllib3
tkinter
Pillow